+++ derive_builder_core/src/builder.rs
@@ -92,8 +92,10 @@ pub struct Builder<'a> {
     pub pattern: BuilderPattern,
     /// Traits to automatically derive on the builder type.
     pub derives: &'a [Path],
-    /// Attributes to include on the builder `struct` declaration and its inherent `impl`.
-    pub attrs: &'a [syn::Attribute],
+    /// Attributes to include on the builder `struct` declaration.
+    pub struct_attrs: &'a [syn::Attribute],
+    /// Attributes to include on the builder's inherent `impl` block.
+    pub impl_attrs: &'a [syn::Attribute],
     /// When true, generate `impl Default for #ident` which calls the `create_empty` inherent method.
     ///
     /// Note that the name of `create_empty` can be overridden; see the `create_empty` field for more.
@@ -170,7 +172,8 @@ impl<'a> ToTokens for Builder<'a> {
                 }
             };
 
-            let attrs = self.attrs;
+            let struct_attrs = self.struct_attrs;
+            let impl_attrs = self.impl_attrs;
 
             let builder_doc_comment = &self.doc_comment;
             let deprecation_notes = &self.deprecation_notes.as_item();
@@ -178,9 +181,12 @@ impl<'a> ToTokens for Builder<'a> {
             #[cfg(not(feature = "clippy"))]
             tokens.append_all(quote!(#[allow(clippy::all)]));
 
+            // struct_attrs MUST come after derive_attr, otherwise attributes for a derived
+            // trait will appear before its derivation. As of rustc 1.59.0 this is a compiler
+            // warning; see https://github.com/rust-lang/rust/issues/79202
             tokens.append_all(quote!(
-                #(#attrs)*
                 #derive_attr
+                #(#struct_attrs)*
                 #builder_doc_comment
                 #builder_vis struct #builder_ident #struct_generics #where_clause {
                     #(#builder_fields)*
@@ -191,7 +197,7 @@ impl<'a> ToTokens for Builder<'a> {
             tokens.append_all(quote!(#[allow(clippy::all)]));
 
             tokens.append_all(quote!(
-                #(#attrs)*
+                #(#impl_attrs)*
                 #[allow(dead_code)]
                 impl #impl_generics #builder_ident #ty_generics #where_clause {
                     #(#functions)*
@@ -330,7 +336,8 @@ macro_rules! default_builder {
             ident: syn::Ident::new("FooBuilder", ::proc_macro2::Span::call_site()),
             pattern: Default::default(),
             derives: &vec![],
-            attrs: &vec![],
+            struct_attrs: &vec![],
+            impl_attrs: &vec![],
             impl_default: true,
             create_empty: syn::Ident::new("create_empty", ::proc_macro2::Span::call_site()),
             generics: None,